Move tty.Buffer to output/outputtest.Buffer
